Added Gallery styles
