+++ src/assets/styles/gallery2.css
@@ -1,0 +1,29 @@
+@media (max-width: 800px) {
+    .gallery-main-boxes{
+        height:15rem;
+    }
+}
+.gallery-main-boxes {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    box-sizing: border-box;
+    background: #0c9a9a;
+    color: #fff;
+    color: #fff;
+    background-size: cover;
+    background-position: center;
+    box-shadow: -2px 2px 10px 0px rgba(68, 68, 68, 0.4);
+    transition: transform 0.3s ease-in-out;
+    cursor: pointer;
+    counter-increment: item-counter;
+    border-radius: 10px;
+    border: 1px solid black;
+}
+.gallery-main-boxes:hover{
+    outline: 0.875rem solid rgba(0, 0, 0, 0.6);
+    opacity:0.9;
+    transition: 300ms;
+}
