+++ src/pages/gallery/index.js
@@ -1,113 +1,55 @@
 import React, { Component } from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faPlusCircle,
-  faChevronRight,
-  faChevronLeft,
-  faTimes,
-} from "@fortawesome/free-solid-svg-icons";
-import "../../assets/styles/gallery.css";
 import Layout from "components/UI/Layout/Layout";
-import axios from "axios";
+import { Link } from "gatsby";
+import "assets/styles/gallery2.css";
 
-export default class ImageContainer extends Component {
+export default class GallerySecondary extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      FullImageCard: false,
-      imageUrl: null,
-      currentIndex: 0,
-      dummy: [],
-    };
-  }
-  async componentDidMount() {
-    console.log("Mount running");
-    await axios
-      .get("https://spec-backend.herokuapp.com/api/gallery/")
-      .then((response) => {
-        this.setState({ dummy: response.data });
-        console.log(this.state.dummy);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+    this.state = {};
   }
 
-  async showImage(img, ind) {
-    await this.setState({
-      FullImageCard: true,
-      imageUrl: img,
-      currentIndex: ind,
-    });
-  }
-  async showPrev(ind) {
-    if (ind > 0) {
-      let url = this.state.dummy[ind - 1].image_url;
-      await this.setState({
-        FullImageCard: true,
-        imageUrl: url,
-        currentIndex: ind - 1,
-      });
-    }
-    console.log(this.state.currentIndex);
-  }
-  async showNext(ind) {
-    if (ind < this.state.dummy.length - 1) {
-      let url = this.state.dummy[ind + 1].image_url;
-      await this.setState({
-        FullImageCard: true,
-        imageUrl: url,
-        currentIndex: ind + 1,
-      });
-    }
-  }
-  async exitButton() {
-    await this.setState({ FullImageCard: false });
-  }
   render() {
     return (
       <Layout>
-        <div className="gallery-color-scheme">
-          <h1>Gallery</h1>
-          {this.state.FullImageCard && (
-            <div id="overlay">
-              <div
-                id="prevButton"
-                className={
-                  this.state.currentIndex == 0 ? "gallery-disabled" : ""
-                }
-                onClick={() => this.showPrev(this.state.currentIndex)}
-              >
-                <FontAwesomeIcon icon={faChevronLeft} />
-              </div>
-              <img src={this.state.imageUrl} />
-              <div
-                id="nextButton"
-                className={
-                  this.state.currentIndex == this.state.dummy.length - 1
-                    ? "gallery-disabled"
-                    : ""
-                }
-                onClick={() => this.showNext(this.state.currentIndex)}
-              >
-                <FontAwesomeIcon icon={faChevronRight} />
-              </div>
-              <div id="exitButton" onClick={() => this.exitButton()}>
-                <FontAwesomeIcon icon={faTimes} />
-              </div>
+          <h1 className="text-5xl font-bold text-center mt-16 sm:text-7xl md:text-8xl">Gallery</h1>
+
+        <div className="mx-10 my-28 xl:mx-28 grid gap-4 xl:gap-8 grid-cols-1 md:grid-cols-2 lg:grid-cols-4 text-2xl font-bold xl:text-3xl">
+          <div className="w-58 h-120 lg:h-120 lg:row-span-2 bg-gray-400 object-center gallery-main-boxes" style={{ backgroundImage: `url('images/gallery/specfest.jpg')` }}>
+            <div className="gallay-mainpage-inner">
+              <Link to="/gallery/FinalGallery/?page=SpecFest">
+                <button className="text-2xl font-bold xl:text-3xl hover:bg-gray-100 hover:bg-opacity-20 text-white font-bold py-2 px-4 rounded">
+                <h3>SPECFEST</h3>
+                </button>
+              </Link>
             </div>
-          )}
-          <div className="image-container-background">
-            <div className="gallery-grid">
-              {this.state.dummy.map((element, index) => (
-                <div
-                  className="gallery-item"
-                  style={{ backgroundImage: `url(${element.image_url})` }}
-                  onClick={() => this.showImage(element.image_url, index)}
-                >
-                  <div className="gallery-item__details">{element.event}</div>
-                </div>
-              ))}
+          </div>
+
+          <div className="w-58 h-60 bg-gray-400 object-scale-down gallery-main-boxes" style={{ backgroundImage:  `url('images/gallery/electrothon.jpg')` }}>
+            <div className="gallay-mainpage-inner">
+              <Link to="/gallery/FinalGallery/?page=Electrothon">
+              <button className="text-2xl font-bold xl:text-3xl hover:bg-gray-100 hover:bg-opacity-20 text-white font-bold py-2 px-4 rounded">
+                <h3>ELECTROTHON</h3>
+                </button>
+              </Link>
+            </div>
+          </div>
+          <div className="w-58 h-60 lg:col-span-2 bg-gray-400 object-cover gallery-main-boxes" style={{ backgroundImage: `url('images/gallery/workshop.jpg')` }}>
+            <div className="gallay-mainpage-inner">
+              <Link to="/gallery/FinalGallery/?page=Workshop">
+              <button className="text-2xl font-bold xl:text-3xl hover:bg-gray-100 hover:bg-opacity-20 text-white font-bold py-2 px-4 rounded">
+                <h3>WORKSHOP</h3>
+                </button>
+              </Link>
+            </div>
+          </div>
+          <div className="w-58 h-60 lg:col-span-3 bg-gray-400  object-cover gallery-main-boxes" style={{ backgroundImage: `url('images/gallery/random.jpg')` }}>
+            <div className="gallay-mainpage-inner">
+              <Link to="/gallery/FinalGallery/?page=random">
+              <button className="text-2xl font-bold xl:text-3xl hover:bg-gray-100 hover:bg-opacity-20 text-white font-bold py-2 px-4 rounded">
+                <h3>RANDOM</h3>
+                </button>
+              </Link>
             </div>
           </div>
         </div>
